WIP support Teensy and maybe AVR
WIP - Added support to be able to use some Teensy boards by PJRC.  I have done some testing with a couple of boards of mine, one based on T3.2 and the other on T3.6.

On the 3.2 it uses only the TX bin of the UART, so it uses the half duplex support of the processor.

On the 3.6 board, I used a couple of buffer chips on RX and TX, with a direction pin.  This allowed the signal to be bumped up from the 3.3v level to 5v and uses the support built into the PJRC Arduino code base to control RS485 like chips with direction pins